requirements.txt
yara-python==4.5.0
py7zr==0.21.0